Break check-inbox.php out of home.php
This is the first of several commits that moves a chunk of display logic out of the main home.php file into a subdirectory provisionally labelled "components" and then uses a php include to reference it in home.php.

The intent is to make home.php easier to work with in future commits by reducing its length and complexity.