Cleanup: rename rightbar class used in homepage
Addresses https://github.com/iftechfoundation/ifdb-suggestion-tracker/issues/218.

In an earlier version of this pull request, I just removed the rules hiding "rightbar" and "game-rightbar"-styled elements from the mobile stylesheet entirely, but I realized that those rules are currently preventing a couple of issues that would otherwise crop up on mobile. like weird overlapping on the "browse" page and putting the game page sidebar before the actual game information itself on the "game details" page.

This commit also restricts the restyling I've been doing to the rightbar class to just the instances used on the refactored homepage, and restores the original styles for rightbar class so that the instances I haven't touched should continue to look the same for now. 

Arguably I should have split and redistributed these changes into prior commits, but this way the homepage (the main focus of this pull request) should look coherent throughout most of the commit history.