Fix nav header on pages that include login-persist.php before util.php
PHP does something weird when you introduce circular dependencies; if A includes B and B includes A, it runs B right up until the point where B includes A, then runs A to completion, and then runs B to completion. That's no good when the dependency is something essential for A to run correctly.

So in this PR we're making util.php depend on session-start.php and ensuring that session-start.php depends on nothing, which guarantees that this will be safe.