Exclude "special" reviews from numMemberReviews
Fixes https://github.com/iftechfoundation/ifdb-suggestion-tracker/issues/261

The change appears in the innermost query, `grouped`, where we've changed the `where` to require that `special is null`, and `hasReview` now simply checks whether `review is not null`, instead of the `ifnull` trickery we were doing before: `ifnull(special, review) is not null AS hasReview`

The `special` reviews are by the "special reviewers" from the `specialreviewers` table. (There are no reviews in the DB with special=3.)

```
select * from specialreviewers;
+----+----------+-----------------+-------------+-----------+-------------------+
| id | code     | name            | displayrank | editorial | requiredprivilege |
+----+----------+-----------------+-------------+-----------+-------------------+
|  1 | author   | From the Author |         100 |         0 | NULL              |
|  2 | bafs     | Baf's Guide     |          10 |         1 | E                 |
|  3 | ifdb     | IFDB Review     |           1 |         1 | E                 |
|  4 | external | External        |          50 |         1 | A                 |
+----+----------+-----------------+-------------+-----------+-------------------+
```

Thus, all of these "special" reviews should be excluded from this view.

I'm honestly not sure what happened with that `ifnull(special, review)` trickery. It was in the query when I originally checked it in, back in https://github.com/iftechfoundation/ifdb/pull/71/commits/1aa9b5c8c1f25c07cec4e987153c65be39685c15 but my plain-text summary of the query doesn't have that trickery; the plain-text summary says:

```
(special is null and review is not null) as hasReview
```

If the query had said that, we wouldn't have had to fix this bug, though we still would have had a subtler bug that the histogram of ratings would have included external reviews, which is probably not what we want here.