Add putific lastversion parameter, to modify listings from other users
Previously, `putific` would only allow you to modify listings if you were the most recent user to edit the listing. But the IFDB web UI has a facility for this. When editing a page, it records the last-known version that you knew about. When you submit, it attempts to merge your changes with the existing document.

The `putific` `lastversion` parameter does the same thing, allowing API users to declare the last version they know about, and allowing the merge to go through normally.

This also requires us to add a `<pageversion>` to the `viewgame` API, so API users can discover the last-known version and make changes based on that.