remove inline event handlers: editcomp (gridform refactor)
This one was especially tricky, because editcomp makes heavy use of gridform, which generates HTML in JS as a string and then injects it with innerHTML. There's no way to attach <script> tags with innerHTML, even if the <script> contains the appropriate nonces.

So, to refactor, I added a new hook for gridform, activateListeners, designed to re-activate event handlers for all of the gridform generated HTML, and moved all event handlers in there.

This was especially painful for comboboxes, because I had to manually re-attach equivalent event listeners on all of the generated comboboxes.