Add a cron job to update embargoed reviews
The `gameRatingsSandbox0` view automatically updates to include reviews whose embargo dates have passed, but we need to make an update to the `reviews` table when a review's embargo date passes in order to update the `gameRatingsSandbox0_mv` materialized view.

Here, we're adding an `embargopastdate` column to `reviews`, which we update when the embargo date has actually passed. That update will refresh the materialized view.

And then we're adding an URL that can be managed by a cron job to update `embargopastdate` for any/all reviews that just passed their embargo date.