Regenerate gameRatingsSandbox0_mv when sandboxing a user
When a user enters/leaves the sandbox, their reviews will need to be updated in the `gameRatingsSandbox0_mv` materialized view.

This code is kinda inefficient, regenerating the entire materialized view from scratch whenever any user is added/removed from the sandbox, but as of today that still only takes a second or so, and users don't get sandboxed or unsandboxed very often.