Fix documentation on how to search by IFID or TUID
The search popup claimed that you could search by "title, IFID, or TUID," but it was a lie, because it was really just running the same old search algorithm that the search box uses, which doesn't honor search by IFID.

Briefly, we made the searchbox honor implicitly searching by TUID, but this broke our fulltext index, so we rolled that back. As a result, to search by IFID or by TUID, you have to use the `ifid:` or `tuid:` operators.

This commit updates the search popup to say "Search for a game by title, `ifid:IFID`, or `tuid:TUID`:" which explains things about as clearly as we can.