Don't delete DEFINER, just set it to the `ifdb` user
mysqldump likes to generate `DEFINER=` directives pointing to `root`. In the past, we could just drop any line containing `DEFINER=`, but now that we're using triggers, mysqldump likes to generate those smack dab in the middle of `CREATE TRIGGER` statements, so we need to keep them.

Instead of deleting them, we're now setting `DEFINER=` to `ifdb`@`%`, which restores cleanly.