viewgame: Standardize rating/review counts next to the histogram
When showing the histogram, we should always show the total number of ratings that factored into the histogram, and we should juxtapose that number with the number of written reviews, because these numbers are confusingly similar ideas.

This fixes https://github.com/iftechfoundation/ifdb-suggestion-tracker/issues/475 where the `?ratings` view would show `$rowcnt` as the number of ratings, which might not match `$ratingCntAvg`.