On mobile, show description above External Links (floating right)
This commit is a bit confusing. We want External Links to `float: right` (instead of `position: absolute`), to ensure that it doesn't overlap the tags table or the footer.

But, to make the floated box appear in the upper-right corner of the page, we have to move the External Links section up to the top of the body of the page. That would show it above the title and coverart (the "main summary").

So, then, on mobile, we set the main body of the page to flexbox, and add `order: -1` to the main summary. This forces the main summary to appear on the page above the External Links on mobile, while allowing the External Links to `float: right` to the upper-right corner of the page on wide displays.