Add a table view (with tag frequency in parentheses) to the showtags page
This is definitely not ready to merge. The intention is to add an option for viewing all the tags in a table, with the frequency of each tag in parentheses. The third column is pushed off to the right, I think because if the length of some of the tags which are actually URLs, but I'm not sure how to fix it.