Add message about the Use Accessible Controls setting
People who might like to use the "Use accessible controls" setting on the settings page may not know it exists. This message is meant to make users aware of that setting, in text that is read by a screen reader but is not visible on the screen.

I have not been able to test it, but it is modeled after the following code. An IFDB user confirmed that the "Yip yip yip" text in the code below is read aloud by a screen reader, even though it is not visible on the screen:

```
<h1>Woof!</h1>
<span style="position:absolute;width:1px;left:-10000px;">Yip yip yip</span>
```