Allow sorting the tags table in alphabetical order
People seemed to like the idea of allowing the table of tags to be sorted in alphabetical order: https://intfiction.org/t/thoughts-on-the-table-of-all-tags-on-ifdb-showtags-page/71482

This adds an option for that.

There are links to show the top 20, 50, and 100 tags in table form--whether in alphabetical order or not--but at the moment, those lines are commented out because I was unsure whether to include them. (It's still possible to list only the top N tags, if you type the limit into the url, but with these lines commented out, that possibility is not advertised).