Make frame black in Edit Link window; make main text white in Edit Review Link window
After changing the background of the Edit Link popup window to black, I saw that there was still a thick border around the text (maybe used to space the text from the edge of the window?). Change this to black to match the color behind the text.

The foreground color in this commit affects the main text in the Edit Review Link window. Make it white to show up against the black background.