Cache CSS & JS immutably when cache busting
We make an effort to request all CSS and JS files with a t= parameter, busting the cache.

But sometimes we can't do that, e.g. when users use a custom stylesheet (e.g. Midnight starts with `@import url("/ifdb.css");
`).

With these settings, when the URL contains a "t=" parameter, we set the file to be publicly cached immutably. When the t parameter is missing, we set it to no-cache, forcing the browser to revalidate its cache on every request.