Cache coverart when version is specified
Whenever we update the coverart, we create a new version of the page. So, when the version is specified, we can immutably cache the coverart image.

I've explicitly added a version parameter to the coverart at the top of the viewgame page and to the search results page. There are more spots to catch, but I think this is a very good start.