Don't re-add stored procedures in full dev environment
Previously, the nightly backup didn't include stored procedures, so we had to manually re-add them.

Now, the nightly backup _does_ contain stored procedures, so we don't have to mess with this any more.

(I'm leaving the patch script in the `sql` directory just in case it comes in handy some day.)