Vertically center Menu toggle in nav bar
Incresing <input> size to 16px (prevent browser zoom on focus) caused the menu bar to grow in height on iOS.

The "Browse" link was vertically centered, but the "Menu" button was aligned to the top.

(I think this is because `<button>` elements kinda refuse to accept `display` directives on their outer display type.)

So, I wrapped the `<button>` in a vertically centered `<div>` and refactored the code to show/hide the menu that way.