Don't autofocus fields in IFDB forms
In many pages on the site, we try to guess which field you want to edit most, and autofocus it.

I find it mildly annoying on desktop; it's super obnoxious on mobile, because it can cause the screen to scroll and the keyboard to popup, obscuring half the screen.

Now, we don't do that.

Fixes #397