Show "No ratings" stars on viewgame only
In #1063, we fixed the viewgame page when there are no ratings yet, to bring back an empty star rating, but this caused empty star ratings to appear all over the site, e.g. home page, off-site reviews, etc.

Now, we only show five empty stars when we opt-in to it (only on the viewgame page).

Fixes #1113