Workaround a bug in Safari `text-wrap: balance`
We were using `text-wrap: balance` here to handle viewports that are small enough to cause the last link to wrap to the next line as a "widow" (or "runt").

`text-wrap: balance` would prevent the widow by moving half of the links to the subsequent line, ensuring that both lines were roughly balanced.

But `text-wrap: balance` seems to be buggy in Safari 18.2; it was causing the first few links to be "orphans" all on their own line. The algorithm seemed to fix itself randomly, e.g. if I uncheck and recheck the box in Safari Dev Tools, or just scroll around a little on the page.

Safari 18.2 doesn't support `text-wrap: pretty`, causing it to fallback to `text-wrap: auto`, which is better than its broken implementation of `text-wrap: balance`.

Fixes #1208