Remove `DISTINCT` from game queries, refix #901
In #902 I added `DISTINCT` to the game search query, to fix a bug where searches by competition could show duplicates, if the same game appeared in multiple divisions.

It turns out that adding `DISTINCT` prevents the query planner from doing simple index scans, e.g. when simply querying for the newest games.

So I've removed `DISTINCT` in this commit, and re-fixed #901 by querying `compgames` with a subquery with a `distinct` clause.

Fixes #1287.