Remove stray `if ($specials)`
In the main branch prior to work on Parsedown, this was a simple `else`. In fe35272b we experimented with making it `else if ($specials & !FixDescMarkdown)`, but when we removed `FixDescMarkdown` in 3b8b9f03 we didn't revert the rest of the line.

This commit makes a simple `fixDesc($txt)` with no `$specials` convert unrecognized entities to `&amp;`, as it used to.