Fix and rename broken Selenium test case
The first Selenium test case didn't work, presumably
because it relied on links in the top bar
that no longer exist when the user is logged out.

I fixed the test case by making it access the webpages
for those links directly. It was checking that a logged-out
user wouldn't be able to access any of them, and I kept
that in, but the test case no longer uses the top bar to navigate.

I renamed the test to "Top Bar When Logged Out" in fitting
with its new purpose.